# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Memory Service",
    "description": "AI agent memory service — ingest conversation turns, extract structured memories, answer recall queries with hybrid BM25 + cosine retrieval.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local Docker" }
  ],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Required only when MEMORY_AUTH_TOKEN is set. Leave blank otherwise."
      }
    },
    "schemas": {
      "Message": {
        "type": "object",
        "required": ["role", "content"],
        "properties": {
          "role": { "type": "string", "enum": ["user", "assistant", "tool"] },
          "content": { "type": "string", "maxLength": 100000 },
          "name": { "type": "string" }
        }
      },
      "Citation": {
        "type": "object",
        "properties": {
          "turn_id": { "type": "string" },
          "score": { "type": "number" },
          "snippet": { "type": "string" }
        }
      },
      "MemoryRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string", "enum": ["fact", "preference", "opinion", "event"] },
          "key": { "type": "string" },
          "value": { "type": "string" },
          "confidence": { "type": "number" },
          "source_session": { "type": "string" },
          "source_turn": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "supersedes": { "type": "string", "nullable": true },
          "active": { "type": "boolean" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "operationId": "getHealth",
        "security": [],
        "responses": {
          "200": {
            "description": "Service is up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "timestamp": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/turns": {
      "post": {
        "summary": "Ingest a conversation turn",
        "description": "Persists the turn, runs two-pass LLM extraction, embeds memories. Synchronous — 201 means memories are immediately queryable.",
        "operationId": "postTurns",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["session_id", "messages", "timestamp"],
                "properties": {
                  "session_id": { "type": "string", "minLength": 1, "maxLength": 500, "example": "sess-abc123" },
                  "user_id": { "type": "string", "maxLength": 500, "nullable": true, "example": "user-42" },
                  "messages": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 100,
                    "items": { "$ref": "#/components/schemas/Message" }
                  },
                  "timestamp": { "type": "string", "format": "date-time", "example": "2024-01-15T10:30:00Z" },
                  "metadata": { "type": "object", "additionalProperties": true, "default": {} }
                }
              },
              "examples": {
                "simple": {
                  "summary": "Single user message",
                  "value": {
                    "session_id": "sess-1",
                    "user_id": "user-1",
                    "messages": [
                      { "role": "user", "content": "I live in Berlin and work at Notion as a product manager. I have a dog named Biscuit." }
                    ],
                    "timestamp": "2024-01-15T10:30:00Z"
                  }
                },
                "multi_turn": {
                  "summary": "Multi-turn conversation",
                  "value": {
                    "session_id": "sess-1",
                    "user_id": "user-1",
                    "messages": [
                      { "role": "user", "content": "I just started at Stripe as an engineer." },
                      { "role": "assistant", "content": "That's exciting! What team are you joining?" },
                      { "role": "user", "content": "I'll be on the payments infrastructure team." }
                    ],
                    "timestamp": "2024-01-15T11:00:00Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Turn persisted and memories extracted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "example": "550e8400-e29b-41d4-a716-446655440000" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "413": { "description": "Payload too large (> 1 MB)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/recall": {
      "post": {
        "summary": "Recall memories for a query",
        "description": "Runs query rewriting → BM25 + cosine → RRF fusion → multi-hop entity extraction. Returns formatted context and citations.",
        "operationId": "postRecall",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query", "session_id"],
                "properties": {
                  "query": { "type": "string", "minLength": 1, "maxLength": 2000, "example": "where does the user live?" },
                  "session_id": { "type": "string", "example": "sess-1" },
                  "user_id": { "type": "string", "nullable": true, "example": "user-1" },
                  "max_tokens": { "type": "integer", "minimum": 64, "maximum": 8192, "default": 1024 }
                }
              },
              "examples": {
                "location": {
                  "summary": "Location query",
                  "value": { "query": "where does the user live", "session_id": "sess-1", "user_id": "user-1", "max_tokens": 1024 }
                },
                "multihop": {
                  "summary": "Multi-hop query",
                  "value": { "query": "what city does the person with the dog named Biscuit live in", "session_id": "sess-1", "user_id": "user-1" }
                },
                "noise": {
                  "summary": "Unrelated query (should return empty)",
                  "value": { "query": "what does the user think about quantum computing", "session_id": "sess-1", "user_id": "user-1" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Context and citations. Empty context means no relevant memories found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "context": { "type": "string", "example": "## Known facts about this user\n- [2024-01-15] location: lives in Berlin" },
                    "citations": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Citation" }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/search": {
      "post": {
        "summary": "Search memories",
        "description": "BM25 + cosine search for user or session scope. Returns raw scored results, not formatted context.",
        "operationId": "postSearch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": { "type": "string", "minLength": 1, "maxLength": 2000, "example": "dog Biscuit" },
                  "session_id": { "type": "string", "nullable": true },
                  "user_id": { "type": "string", "nullable": true, "example": "user-1" },
                  "limit": { "type": "integer", "minimum": 1, "maximum": 50, "default": 10 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "content": { "type": "string" },
                          "score": { "type": "number" },
                          "session_id": { "type": "string" },
                          "timestamp": { "type": "string" },
                          "metadata": { "type": "object" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}/memories": {
      "get": {
        "summary": "List all memories for a user",
        "description": "Returns ALL memories including inactive (superseded) ones. Active=true is the current fact; active=false is historical.",
        "operationId": "getUserMemories",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" }, "example": "user-1" }
        ],
        "responses": {
          "200": {
            "description": "All memories for this user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memories": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/MemoryRecord" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}": {
      "delete": {
        "summary": "Delete a session",
        "description": "Deletes all turns, memories, and embeddings for this session.",
        "operationId": "deleteSession",
        "parameters": [
          { "name": "sessionId", "in": "path", "required": true, "schema": { "type": "string" }, "example": "sess-1" }
        ],
        "responses": {
          "204": { "description": "Session deleted" }
        }
      }
    },
    "/users/{userId}": {
      "delete": {
        "summary": "Delete a user",
        "description": "Deletes all data for this user: turns, memories, embeddings.",
        "operationId": "deleteUser",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" }, "example": "user-1" }
        ],
        "responses": {
          "204": { "description": "User data deleted" }
        }
      }
    }
  }
}